+++ src/api/Program.cs
@@ -144,9 +144,11 @@ builder.Services
     .LoadFromConfig(builder.Configuration.GetSection("ReverseProxy"));
 
 builder.Services.AddScoped<IBoreholePermissionService, BoreholePermissionService>();
+builder.Services.AddScoped<IFilterService, FilterService>();
 builder.Services.AddSingleton(TimeProvider.System);
 builder.Services.AddSingleton<IMaintenanceTask, LocationMigrationTask>();
 builder.Services.AddSingleton<IMaintenanceTask, CoordinateMigrationTask>();
+builder.Services.AddSingleton<IMaintenanceTask, UserMergeTask>();
 builder.Services.AddSingleton<MaintenanceTaskService>();
 
 builder.Services